# docs/swagger.yaml
basePath: /
definitions:
...
  server.swaggerCreateHospitalRequest:
    properties:
      name:
        type: string
    type: object
  server.swaggerCreateNoteRequest:
    properties:
      appointment_id:
        type: string
      content:
        type: string
      hospital_id:
        type: string
      patient_id:
        type: string
    type: object
...
  server.swaggerHospitalResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  server.swaggerHospitalStatsResponse:
    properties:
      active_patients:
        type: integer
      prescription_count:
        type: integer
      total_appointments:
        type: integer
    type: object
...
  server.swaggerNoteResponse:
    properties:
      appointment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  server.swaggerStartDictationRequest:
    properties:
      appointment_id:
        type: string
      hospital_id:
        type: string
    type: object
...
  server.updateNoteRequest:
    properties:
      content:
        type: string
      note_id:
        type: string
    required:
    - content
    - note_id
    type: object
host: localhost:8080
info:
  contact:
    email: support@clinicwise.io
    name: Clinic Wise Support
  description: REST API for the Clinic Wise healthcare management platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Clinic Wise API
  version: "1.0"
paths:
...
  /hospitals:
    post:
      consumes:
      - application/json
      description: Creates a new hospital. Requires admin role.
      parameters:
      - description: Create hospital payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerCreateHospitalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerHospitalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a hospital
      tags:
      - Hospitals
  /hospitals/{id}/stats:
    get:
      description: Returns total appointments, active patients, and prescription count
        for a hospital. Requires admin role.
      parameters:
      - description: Hospital ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerHospitalStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Hospital statistics
      tags:
      - Hospitals
  /notes:
    patch:
      consumes:
      - application/json
      description: Updates the content of an existing note. Requires doctor role.
      parameters:
      - description: Update note payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.updateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Creates a clinical note for an appointment. Requires doctor role.
      parameters:
      - description: Create note payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerCreateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerNoteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a note
      tags:
      - Notes
  /notes/{id}:
    get:
      description: Returns a single note by ID. Requires doctor or patient role and
        ownership.
      parameters:
      - description: Note ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerNoteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a note
      tags:
      - Notes
  /notes/appointment/{id}:
    get:
      description: Returns all notes for a given appointment. Requires doctor or patient
        role and ownership.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerNoteResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notes for an appointment
      tags:
      - Notes
  /notes/dictation:
    get:
      description: Upgrades the connection to a WebSocket for real-time voice dictation.
        Requires doctor role.
      parameters:
      - description: Dictation request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerStartDictationRequest'
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start voice dictation (WebSocket)
      tags:
      - Notes
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"